Conflicts:
	trobo/src/main/java/com/allstate/trobo/controller/TripSheetController.java
	trobo/src/main/java/com/allstate/trobo/service/impl/TripSheetServiceImpl.java
	trobo/src/main/resources/schema.sql
	trobo/src/main/webapp/resources/js/sb-admin-2.js
	trobo/target/classes/schema.sql

+++ trobo/src/main/java/com/allstate/trobo/dao/impl/JdbcTripsheetRepository.java
@@ -59,8 +59,8 @@ public class JdbcTripsheetRepository implements TripSheetRepository {
 			Address address = new Address();
 			address.setId(rs.getLong("id"));
 			address.setAddressLine(rs.getString("address_line"));
-			address.setLongitude(rs.getBigDecimal("longitude"));
-			address.setLatitude(rs.getBigDecimal("latitude"));
+			address.setLongitude(rs.getDouble("longitude"));
+			address.setLatitude(rs.getDouble("latitude"));
 			pickUpPoint.setAddress(address);
 			pickUpPoint.setNumberOfEmployees(rs.getInt("empCount"));
 			return pickUpPoint;
